refactor(web): rename resume interfaces to schema
Change references from 'interfaces' to 'schema' directory in resume
module to align with updated project structure and improve code
organization.
