+++ apps/web/app/routes/resume/languages.tsx
@@ -1,6 +1,6 @@
 import { T } from '@suddenly-giovanni/ui/components/typography/typography.tsx'
 import type { ReactElement } from 'react'
-import type { Language } from '~/routes/resume/interfaces/language.ts'
+import type { Language } from '~/routes/resume/schema/language.ts'
 
 export function Languages({
 	languages,
