+++ apps/web/app/routes/resume/route.tsx
@@ -9,7 +9,7 @@ import resumeAssetUrl from '@suddenly-giovanni/resume/resume.json?raw'
 import { clsx } from '@suddenly-giovanni/ui/lib/utils.ts'
 import type { ReactElement } from 'react'
 import { routesRecord } from '~/routes-record.ts'
-import { Resume as ResumeSchema } from '~/routes/resume/interfaces'
+import { Resume as ResumeSchema } from 'app/routes/resume/schema'
 import { Contacts } from './contacts.tsx'
 import { Education } from './education.tsx'
 import { Experiences } from './experiences.tsx'
