+++ apps/web/app/routes/resume/schema/basics.ts
@@ -1,6 +1,6 @@
 // biome-ignore lint/nursery/noNamespaceImport: this is how we import from schema
 import * as S from '@effect/schema/Schema'
-import { Phone } from '~/routes/resume/interfaces/phone.ts'
+import { Phone } from './phone.ts'
 import { Email } from './email.ts'
 import { Location } from './location.ts'
 import { Profile } from './profile.ts'
